refactor ts.carma module, sim and fit  are in two mod now
