+++ src/eztao/ts/__init__.py
@@ -1,2 +1,3 @@
 from .utils import *
-from .carma import *
+from .carma_fit import *
+from .carma_sim import *
